requirements.txt
Pillow <10.0.0
easyocr>=1.7.0
numpy>=1.19.0
mcp>=1.0.0 